Split workspace activity surface builder
